Move SessionParameters inside messaging configurations
