+++ src/lib/dnssd/BUILD.gn
@@ -27,7 +27,7 @@ static_library("dnssd") {
     "${chip_root}/src/crypto",
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/support",
-    "${chip_root}/src/messaging:messaging_mrp_config",
+    "${chip_root}/src/messaging:configurations",
   ]
 
   sources = [
