+++ src/messaging/BUILD.gn
@@ -33,8 +33,11 @@ if (chip_config_resolve_peer_on_first_transmit_failure != "") {
   defines += [ "CHIP_CONFIG_RESOLVE_PEER_ON_FIRST_TRANSMIT_FAILURE=${chip_config_resolve_peer_on_first_transmit_failure}" ]
 }
 
-source_set("messaging_mrp_config") {
-  sources = [ "ReliableMessageProtocolConfig.h" ]
+source_set("configurations") {
+  sources = [
+    "SessionParameters.h",
+    "ReliableMessageProtocolConfig.h",
+  ]
 
   public_deps = [ "${chip_root}/src/system" ]
 }
@@ -67,7 +70,7 @@ static_library("messaging") {
   cflags = [ "-Wconversion" ]
 
   public_deps = [
-    ":messaging_mrp_config",
+    ":configurations",
     "${chip_root}/src/app/icd/server:icd-server-config",
     "${chip_root}/src/crypto",
     "${chip_root}/src/inet",
