+++ src/messaging/SessionParameters.h
