+++ src/protocols/secure_channel/PairingSession.h
@@ -30,7 +30,7 @@
 #include <messaging/ExchangeContext.h>
 #include <protocols/secure_channel/Constants.h>
 #include <protocols/secure_channel/SessionEstablishmentDelegate.h>
-#include <protocols/secure_channel/SessionParameters.h>
+#include <messaging/SessionParameters.h>
 #include <protocols/secure_channel/StatusReport.h>
 #include <transport/CryptoContext.h>
 #include <transport/SecureSession.h>
